+++ src/admin/dashboard/components/Dashboard.tsx
@@ -13,6 +13,7 @@ import AdminDashboard from '../pages/dashboard';
 import VendorsPage from '../pages/vendors';
 import ReverseWithdrawalPage from '../pages/reverse-withdrawal';
 import ReverseWithdrawalTransactionPage from 'admin/dashboard/pages/reverse-withdrawal/ReverseWithdrawalTransaction';
+import ChangelogPage from '../pages/changelog';
 
 export type DokanAdminRoute = {
     id: string;
@@ -72,6 +73,11 @@ const getAdminRoutes = () => {
             id: 'reverse-withdrawal-store',
             element: <ReverseWithdrawalTransactionPage />,
             path: '/reverse-withdrawal/store/:id',
+        },
+        {
+            id: 'changelog',
+            element: <ChangelogPage />,
+            path: '/changelog',
         },
     ];